requirements.txt
altgraph==0.17.4
future==1.0.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pytesseract==0.3.13
pywin32-ctypes==0.2.3
setuptools==75.8.0
tk==0.1.0
